refactor: 컨트롤러의 행동 클래스를 phase 패키지로 이동
