+++ src/main/java/controller/phase/GamePhase.java
@@ -1,4 +1,6 @@
-package controller;
+package controller.phase;
+
+import controller.GameContext;
 
 public interface GamePhase {
