+++ src/main/java/controller/phase/ShowBettingResultPhase.java
@@ -1,5 +1,6 @@
-package controller;
+package controller.phase;
 
+import controller.GameContext;
 import model.BettingResult;
 import view.OutputView;
